% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_watercolour}
\alias{scale_fill_watercolour}
\title{Setup fill palette for ggplot2}
\usage{
scale_fill_watercolour(..., palette = "zorn", type = "discrete", alpha = 1)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'watercolour_palettes' list}

\item{type}{A string representing the palette type, should be one of
discrete, continuous or diverging}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{discrete}{whether to use a discrete colour palette}
}
\description{
Setup fill palette for ggplot2
}
