% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_palette}
\alias{show_palette}
\title{Show a all palette types for a named palette}
\usage{
show_palette(name)
}
\arguments{
\item{name}{The name of a watercolour palette}
}
\description{
Display all the type variations of a single palette to see whether it meets
your needs.
}
\examples{
show_palette("richmond")
}
