% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{watercolour}
\alias{watercolour}
\title{Watercolour Colour Palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
watercolour
}
\description{
Colour palettes derived from paintings from the watercolour artists both
contemporary and masters. Inspired by the \href{https://github.com/ropenscilabs/ochRe}{ochRe}
package and the \href{https://github.com/EdwinTh/dutchmasters}{dutchmasters}
package.
}
\details{
Available palettes:

Anders Zorn - Sommarnöje (1886) - https://commons.wikimedia.org/wiki/File:Sommarn\%C3\%B6je_(1886),_akvarell_av_Anders_Zorn.jpg

Fidelia Bridges - Calla Lilly (1875) - https://commons.wikimedia.org/wiki/File:Brooklyn_Museum_-\emph{Calla_Lily}-\emph{Fidelia_Bridges}-_overall.jpg

Dorothy (Dolla) Richmond - Mount Egmont (1929) - https://commons.wikimedia.org/wiki/File:D._K.\emph{Richmond}-\emph{Mount_Egmont}-_Google_Art_Project.jpg
}
\examples{
viz_palette(watercolour$zorn$discrete)
}
\keyword{datasets}
