% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{watercolour_pal}
\alias{watercolour_pal}
\title{watercolour palette with ramped colours}
\usage{
watercolour_pal(palette = "zorn", type = "discrete", alpha = 1)
}
\arguments{
\item{palette}{Choose from 'watercolour_palettes' list}

\item{type}{A string representing the palette type, should be one of
discrete, continuous or diverging}

\item{alpha}{transparency}
}
\description{
watercolour palette with ramped colours
}
