% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{watercolours}
\alias{watercolours}
\title{Watercolours Colour Palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
watercolours
}
\description{
Colour palettes derived from paintings from the watercolour artists both
contemporary and masters. Inspired by the \href{https://github.com/ropenscilabs/ochRe}{ochRe}
package and the \href{https://github.com/EdwinTh/dutchmasters}{dutchmasters}
package.
}
\details{
Available palettes:

Frank Webb - Unknown artwork
}
\examples{
viz_palette(watercolours$frankwebb)
}
\keyword{datasets}
