% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{watercolours_pal}
\alias{watercolours_pal}
\title{watercolours palette with ramped colours}
\usage{
watercolours_pal(palette = "frankwebb", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'watercolours_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
watercolours palette with ramped colours
}
