% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_watercolour}
\alias{scale_color_watercolour}
\alias{scale_colour_watercolour}
\title{Setup colour palette for ggplot2}
\usage{
scale_color_watercolour(..., palette = "zorn", type = "discrete", alpha = 1)

scale_colour_watercolour(..., palette = "zorn", type = "discrete", alpha = 1)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'watercolour_palettes' list}

\item{type}{A string representing the palette type, should be one of
discrete, continuous or diverging}

\item{alpha}{transparency}
}
\description{
Setup colour palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl)), size = 4) +
  scale_colour_watercolour()
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hp)) +
  scale_colour_watercolour(palette="zorn", type = "continuous")
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_colour_watercolour(palette="richmond")
}
