% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{watercolour_pal}
\alias{watercolour_pal}
\title{watercolour palette with ramped colours}
\usage{
watercolour_pal(palette = "frankwebb", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'watercolour_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
watercolour palette with ramped colours
}
